[minor] created new screens to hold movie list and movie detail
