+++ app/src/main/java/com/devspacecinenow/MovieDetailScreen.kt
@@ -1,0 +1,28 @@
+package com.devspacecinenow
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun MovieDetailScreen() {
+    MovieDetailContent()
+}
+
+@Composable
+private fun MovieDetailContent() {
+    Column(
+        modifier = Modifier.fillMaxSize()
+    ) {
+        Text(text = "Movie detail Screen")
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun MovieDetailPreview() {
+    MovieDetailContent()
+}
